+++ packages/async_job/test/cleanup_rakes_test.dart
@@ -3,7 +3,7 @@ library;
 
 import 'dart:async';
 
-import 'package:async_job/async_job.dart';
+import 'package:async_job/engine.dart';
 import 'package:test/test.dart';
 
 import 'support/error_observer.dart';