+++ packages/async_job/test/support/probe_job.dart
@@ -1,4 +1,4 @@
-import 'package:async_job/async_job.dart';
+import 'package:async_job/engine.dart';
 
 import 'cancel_reason.dart';